Add Lwt_eio.t token to ensure library is initialised
`with_event_loop` now passes a `Lwt_eio.Token.t` to its callback.

If a library needs Lwt_eio to be running, it can request this token to
remind the application to initialise it. This is most useful if a
library has a main entry point, or some configuration value that has to
be constructed before the library is used.

This type should be treated as abstract, but it is defined here as
`private ()` to avoid breaking existing code. It's in a `Token` module
in case someone does `open Lwt_eio`.